Allow override of gem_provider
Without this change, users are not able to set their own gem provider.
This is a problem in environment where the default params logic does not
yield the correct provider, such as in the case of using puppetserver.
This change surfaces the gem_provider param to the autosign class so
that the references to autosign::gem_provider work as expected.